Adjusted a bit of code to work cleaner.
Made the mappings make more sense for the HID edge of things (since it
doesn't QUITE match the libusb RE-ed behaviors...) and removed some
"junk" from the code fromthe old version of things.  We're still getting
spurious events from the HID edge on things so I've blocked that
behavior off as best as I could so we're not pushing junk back to the
caller.  We're also still getting the input class driver edge event
keypresses from the keyboards- so there's still a grab input aspect or a
shut the noise off so I can control when it goes to /dev/input to be
done for this.  It's definitely something worth looking at as a very
early alpha now.