Added /dev/input event blocking.  Made reset() official.
Updated codebase to not do resets under all cases since I couldn't
find a good, safe, non-portable way to force a reset of all devices.

(It's all good...we're only really needing resets if we're bringing
 a daemon down and back up again- and we expect to have that so solid
 that this is a non-issue...)
 
Put in /dev/input event blocking.  This means that any of these 
keyboards that expose an event interface for /dev/input for the
G-Keys (G15, G510s, etc...) will now be ignored by Linux for the
purposes of junk going to the input subsystem.  We want/need to
control the events so that we're sending them by application in
a controlled manner rather than by way of the OS layer, which 
doesn't QUITE get it right.