General updates.
- Updated README.md to reflect the new roadmap plans.  Bigger picture
planned with this, even if Gnome15 stays partly or mostly intact.  The
"daemon" mode may be exposed as a CUSE driver to be "consistent" with
the Logitech "kernel" drivers out and normalize this work.  There's a
bigger picture with more than the original G15 interface that I'd like
to try to account for as "wonderkeys", which handles LED backlight
setting as a simple command (without the driver layer...since it's
likely to not be needed there...), a background layer that cleanly maps
the G-Keys and is configurable for display, RGB backlight, etc., and
something that exposes a similar interface for high-function keyboards
like the G15/G19/Razer Deathstalker Ultimate.
- Corrected capabilities definition of the G710+.  It's a backlight only
keyboard.
- Cleaned up some things that would compile correct, but Eclipse flagged
as a bug (rightly so...  No break; statement in the default on a
switch...)