Added a C interface variant to have for Gnome15 or similar
Re-worked things slightly to provide an C exportable interface to
the framework.  There's an assumption made by this framework
which will precipitate some need for minor adjustments on the
Gnome15 side as it's first consumer.  Joysticks are presumed
to be part of a device EVENT, which can have keys AND joystick
values off of the G13 Keypad.  The original design (incl. the
"direct" G15 driver) presumes Joysticks are separately pollable
(Which they NEVER WERE...)  State for the Joysticks needs to be
held in the consumer, which we preaume in this design.  The
Gnome15 "driver" code doesn't do this, soo....