Updated README.md
Updated my readme to reflect the current state of affairs.  In reality, we're real close to a bolt-on
of this to a fork of Gnome15 for a PoC that this will work and work cleanly for damned near any
G15 class of device without hand-holding, without risking being locked out of the keyboard if
there's a lockup of Gnome15 (which is possible...it happened enough that it wasn't stable on
my setup for the longest time...), and without needing to blacklist HID drivers and force others
which may/may not work with things correctly (For example, G510/G510s isn't really supported
gracefully with anything other than the direct driver, libg15 (sort-of).  The kernel modules were
a joke for those.

This is intended to unify the bulk of G15 class device support under ONE unifying, clean, stable 
interface that is the same basic one that the Logitech Windows code uses.